+++ internal/controller/httprest/api/metrics.go
@@ -7,6 +7,7 @@ import (
 	"sort"
 	"strconv"
 
+	"github.com/crazylazyowl/metrics-tpl/internal/controller/httprest/middleware"
 	"github.com/crazylazyowl/metrics-tpl/internal/usecase/metrics"
 
 	"github.com/go-chi/chi/v5"
@@ -18,6 +19,33 @@ type MetricsAPI struct {
 
 func NewMetricsAPI(metrics *metrics.Usecase) *MetricsAPI {
 	return &MetricsAPI{metrics: metrics}
+}
+func NewMetricsRouter(metrics *metrics.Usecase) http.Handler {
+	api := NewMetricsAPI(metrics)
+
+	r := chi.NewRouter()
+
+	r.Group(func(r chi.Router) {
+		r.Use(middleware.Compress)
+		r.Get("/", api.GetMetrics)
+	})
+
+	// text/plain routes
+	r.Group(func(r chi.Router) {
+		r.Get("/value/{type}/{metric}", api.GetMetric)
+		r.Post("/update/{type}/{metric}/{value}", api.UpdateMetric)
+	})
+
+	// json routes
+	r.Group(func(r chi.Router) {
+		r.Use(middleware.JSONContentType)
+		r.Use(middleware.Compress)
+		r.Use(middleware.Decompress)
+		r.Post("/value/", api.GetMetricJSON)
+		r.Post("/update/", api.UpdateMetricJSON)
+	})
+
+	return r
 }
 
 func (api *MetricsAPI) GetMetrics(w http.ResponseWriter, r *http.Request) {
